Separate out base classes into their own modules
Note: MicroPython does not support __all__: https://docs.micropython.org/en/v1.14/genrst/core_language.html#import
